add team photo
Team photo is directly below title as requested in canvas. Thank you @MyNameIsCalvinDavis for making this team photo! for the caption, I think that putting the names corresponding to the photo would be appropriate (matching names from left to right). I am sorry as I have not matched names and faces yet.