Fixed all bugs, cleaned up code
One issue still exists, maybe its not a problem, but when you load ski.json and press the shorten trip button, it optimizes it seemingly correctly, and when done a second time it optimizes THAT distance. Maybe this is related to the fact that given two neighbors the same distance away, a node will always pick the one it iterates over first. Optimizing it a second time changes the order, and so re-shuffles everything, showing a different map.